% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir}
\alias{dir}
\alias{dir_size}
\alias{dir_cleanup}
\title{Directory helpers: individual directories}
\usage{
dir_size(.sink, .folder = NULL, ..., .unit = c("MB", "GB", "TB"))

dir_cleanup(.sink)
}
\arguments{
\item{.sink}{A \code{character} string that defines the directory in which files are located.}

\item{.folder}{(optional) A \code{character} string that defines the name of a sub-folder for which to list files (via \code{\link[=files_list_numbered_types]{files_list_numbered_types()}}) or summarise file sizes (via \code{\link[=dir_size]{dir_size()}}).}

\item{...}{is a placeholder for additional arguments passed to \code{\link[=list.files]{list.files()}}, such as \code{pattern}, excluding \code{full.names}.}

\item{.unit}{For \code{\link[=dir_size]{dir_size()}}, \code{.unit} is a \code{character} string that defines the units of the output (\code{MB}, \code{GB}, \code{TB}).}
}
\value{
\itemize{
\item \code{\link[=dir_size]{dir_size()}} returns a \code{double};
\item \code{\link[=dir_cleanup]{dir_cleanup()}} returns \code{invisible(NULL)}
}
}
\description{
These are simple directory helpers.
}
\details{
\itemize{
\item \code{\link[=dir_size]{dir_size()}} calculates the total size of files in a directory.
\item \code{\link[=dir_cleanup]{dir_cleanup()}} deletes temporary files and or directories recursively;
}
}
\examples{
# Set up example
temp <- tempdir()
sink <- file.path(temp, "proj.file")
dir.create(sink, recursive = TRUE)
write.table("", file = file.path(sink, "1.csv"))
write.table("", file = file.path(sink, "2.csv"))

# Use `file_size()` to summarise file sizes
dir_size(temp, "proj.file")
dir_size(temp, "proj.file", .unit = "GB")
dir_size(temp, "proj.file", .unit = "TB")

# Use `dir_cleanup()` to cleanup directories
dir_cleanup(sink)
stopifnot(isFALSE(dir.exists(sink)))
}
\author{
Edward Lavender
}
