% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{dirs}
\alias{dirs}
\alias{dirs.create}
\alias{dirs.copy}
\title{Directory helpers: multiple directories}
\usage{
dirs.create(paths)

dirs.copy(from, to, cl)
}
\arguments{
\item{paths}{For \code{\link[=dirs.create]{dirs.create()}}, \code{paths} is \code{character} vector of directories.}

\item{from, to, cl}{Arguments for \code{\link[=dirs.copy]{dirs.copy()}}.
\itemize{
\item \code{from} is a \code{character} vector of directories to copy.
\item \code{to} is a \code{character} vector of destination directories.
\item \code{cl} is a cluster argument passed to \code{\link[pbapply:pbapply]{pbapply::pblapply()}}.
}}
}
\value{
\itemize{
\item \code{\link[=dirs.create]{dirs.create()}} returns \code{invisible(NULL)};
\item \code{\link[=dirs.copy]{dirs.copy()}} returns a \code{\link{data.table}} with \code{from} and \code{to} columns plus \code{success} (\code{TRUE}/\code{FALSE});
}
}
\description{
These functions create/copy directories.
}
\details{
\itemize{
\item \code{\link[=dirs.create]{dirs.create()}} is a simple wrapper around \code{\link[=dir.create]{dir.create()}} that creates directories. Directories that already exist are skipped.
\item \code{\link[=dirs.copy]{dirs.copy()}} copies directories (including their contents) from \code{from} to \code{to}.
}
}
\author{
Edward Lavender
}
