% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file}
\alias{file}
\alias{file_path_validated}
\title{File helpers: individual files}
\usage{
file_path_validated(...)
}
\arguments{
\item{...}{\code{character} vectors passed to \code{\link[=file.path]{file.path()}}.}
}
\value{
\itemize{
\item \code{\link[=file_path_validated]{file_path_validated()}} returns a \code{character} string that defines the file path.
}
}
\description{
These are simple file helpers.
}
\details{
\itemize{
\item \code{\link[=file_path_validated]{file_path_validated()}} is a simple wrapper for \code{\link[=file.path]{file.path()}} constructs a file path and verifies that it exists.
}
}
\examples{
# Set up example
temp <- tempdir()
sink <- file.path(temp, "proj.file")
dir.create(sink, recursive = TRUE)
write.table("", file = file.path(sink, "1.csv"))
write.table("", file = file.path(sink, "2.csv"))

# Use `file_path_validated()` to construct & validate file paths
file_path_validated(temp, "proj.file")

dir_cleanup(sink)
}
\author{
Edward Lavender
}
