% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\alias{files_source}
\alias{files_source_r}
\alias{files_list_numbered_types}
\title{File helpers: file vectors}
\usage{
files_source(.sink, .folder = NULL, recursive = TRUE, ...)

files_source_r(
  .sink,
  .folder = NULL,
  pattern = "\\\\.R$",
  recursive = TRUE,
  ...
)

files_list_numbered_types(.sink, .folder = NULL, ...)
}
\arguments{
\item{.sink, .folder}{Character strings that specify the directory and sub-folder within which to list files.}

\item{..., pattern, recursive}{Arguments passed to \code{\link[=list.files]{list.files()}}.}
}
\description{
These are simple file helpers for dealing with multiple files.
}
\details{
\itemize{
\item \code{\link{files_source}} sources files.
\item \code{\link[=files_source_r]{files_source_r()}} sources \code{R} scripts.
\item \code{\link[=files_list_numbered_types]{files_list_numbered_types()}} creates an ordered \code{list} of numbered files. This function expects files to be named \verb{1.\{.ext\}}, \verb{2.\{.ext\}}, ..., \verb{N.\{.ext\}}. All listed files must share the same file extension.
}
}
\examples{
# Set up example
temp <- tempdir()
sink <- file.path(temp, "proj.file")
dir.create(sink, recursive = TRUE)
write.table("", file = file.path(sink, "1.csv"))
write.table("", file = file.path(sink, "2.csv"))

# Use `file_list_numbered_types()` to list files
files_list_numbered_types(temp, "proj.file", pattern = "*.csv$")

dir_cleanup(sink)
}
\author{
Edward Lavender
}
