% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{heres}
\alias{heres}
\alias{here_data_raw}
\alias{here_data}
\alias{here_r}
\alias{here_src}
\alias{here_dev}
\alias{here_fig}
\alias{here_doc}
\title{Directory helpers: paths}
\usage{
here_data_raw(...)

here_data(...)

here_r(...)

here_src(...)

here_dev(...)

here_fig(...)

here_doc(...)
}
\arguments{
\item{...}{Character vectors (see \code{\link[here:here]{here::here()}}).}
}
\value{
The functions return a file path.
}
\description{
These functions are simple wrappers for \code{\link[here:here]{here::here()}} for standard RStudio Project directories.
}
\examples{
here_data_raw()
here_data()
here_r()
here_dev()
here_fig()
here_doc()
}
\author{
Edward Lavender
}
