% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.file-package.R
\docType{package}
\name{proj.file}
\alias{proj.file-package}
\alias{proj.file}
\title{\code{\link{proj.file}}}
\description{
Scientific project development and management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/proj.file/}
  \item Report bugs at \url{https://github.com/edwardlavender/proj.files/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
